backend/services/task_processing.py
import asyncio
import hashlib
import json
import logging
import os
import tempfile
from typing import Dict, Any, Optional

import torch
import tensorflow as tf
import numpy as np
from pydantic import BaseModel

class TaskProcessingService:
    def __init__(self):
        self.logger = logging.getLogger(__name__)
        self.supported_frameworks = {
            'pytorch': self._process_pytorch_task,
            'tensorflow': self._process_tensorflow_task
        }

    async def process_ai_task(self, task_data: Dict[Any, Any]) -> Dict[str, Any]:
        """
        Process an AI computation task with multi-framework support
        
        Expected task_data structure:
        {
            'framework': 'pytorch' or 'tensorflow',
            'task_type': 'training' or 'inference',
            'model_config': {...},
            'input_data': [...],
            'hyperparameters': {...}
        }
        """
        try:
            # Validate task data
            self._validate_task_data(task_data)

            # Select processing method based on framework
            framework = task_data.get('framework', '').lower()
            if framework not in self.supported_frameworks:
                raise ValueError(f"Unsupported framework: {framework}")

            # Process task using framework-specific method
            processing_method = self.supported_frameworks[framework]
            result = await processing_method(task_data)

            # Generate result hash for verification
            result_hash = self._generate_result_hash(result)

            return {
                'result': result,
                'result_hash': result_hash,
                'status': 'completed'
            }

        except Exception as e:
            self.logger.error(f"Task processing error: {str(e)}")
            return {
                'result': None,
                'result_hash': None,
                'status': 'failed',
                'error': str(e)
            }

    def _validate_task_data(self, task_data: Dict[Any, Any]):
        """
        Validate input task data for completeness and security
        """
        required_keys = ['framework', 'task_type', 'model_config', 'input_data']
        for key in required_keys:
            if key not in task_data:
                raise ValueError(f"Missing required key: {key}")

        # Add additional validation rules
        if len(task_data['input_data']) == 0:
            raise ValueError("Input data cannot be empty")

        # Limit input data size
        max_input_size = 100 * 1024 * 1024  # 100 MB
        input_size = len(json.dumps(task_data['input_data']))
        if input_size > max_input_size:
            raise ValueError("Input data exceeds maximum allowed size")

    async def _process_pytorch_task(self, task_data: Dict[Any, Any]) -> Any:
        """
        Process AI tasks using PyTorch
        """
        task_type = task_data.get('task_type')
        model_config = task_data.get('model_config')
        input_data = task_data.get('input_data')
        hyperparameters = task_data.get('hyperparameters', {})

        # Convert input data to PyTorch tensor
        input_tensor = torch.tensor(input_data)

        if task_type == 'training':
            # Simulate model training
            model = self._create_pytorch_model(model_config)
            loss_fn = torch.nn.MSELoss()
            optimizer = torch.optim.Adam(model.parameters(), lr=hyperparameters.get('learning_rate', 0.001))

            # Training loop
            epochs = hyperparameters.get('epochs', 10)
            for _ in range(epochs):
                optimizer.zero_grad()
                output = model(input_tensor)
                loss = loss_fn(output, torch.zeros_like(output))
                loss.backward()
                optimizer.step()

            return model.state_dict()

        elif task_type == 'inference':
            # Perform inference
            model = self._create_pytorch_model(model_config)
            with torch.no_grad():
                result = model(input_tensor)
            return result.numpy().tolist()

        else:
            raise ValueError(f"Unsupported PyTorch task type: {task_type}")

    async def _process_tensorflow_task(self, task_data: Dict[Any, Any]) -> Any:
        """
        Process AI tasks using TensorFlow
        """
        task_type = task_data.get('task_type')
        model_config = task_data.get('model_config')
        input_data = task_data.get('input_data')
        hyperparameters = task_data.get('hyperparameters', {})

        # Convert input data to TensorFlow tensor
        input_tensor = tf.convert_to_tensor(input_data)

        if task_type == 'training':
            # Simulate model training
            model = self._create_tensorflow_model(model_config)
            model.compile(
                optimizer=tf.keras.optimizers.Adam(learning_rate=hyperparameters.get('learning_rate', 0.001)),
                loss='mse'
            )

            # Training
            epochs = hyperparameters.get('epochs', 10)
            model.fit(input_tensor, tf.zeros_like(input_tensor), epochs=epochs, verbose=0)

            return {k: v.numpy().tolist() for k, v in model.get_weights()}

        elif task_type == 'inference':
            # Perform inference
            model = self._create_tensorflow_model(model_config)
            result = model.predict(input_tensor)
            return result.tolist()

        else:
            raise ValueError(f"Unsupported TensorFlow task type: {task_type}")

    def _create_pytorch_model(self, model_config: Dict[str, Any]) -> torch.nn.Module:
        """
        Create a PyTorch model based on configuration
        """
        # Simplified model creation - expand based on actual requirements
        model_type = model_config.get('type', 'linear')
        input_size = model_config.get('input_size', 10)
        output_size = model_config.get('output_size', 1)

        if model_type == 'linear':
            return torch.nn.Sequential(
                torch.nn.Linear(input_size, 64),
                torch.nn.ReLU(),
                torch.nn.Linear(64, output_size)
            )
        else:
            raise ValueError(f"Unsupported PyTorch model type: {model_type}")

    def _create_tensorflow_model(self, model_config: Dict[str, Any]) -> tf.keras.Model:
        """
        Create a TensorFlow model based on configuration
        """
        # Simplified model creation - expand based on actual requirements
        model_type = model_config.get('type', 'sequential')
        input_size = model_config.get('input_size', 10)
        output_size = model_config.get('output_size', 1)

        if model_type == 'sequential':
            return tf.keras.Sequential([
                tf.keras.layers.Dense(64, activation='relu', input_shape=(input_size,)),
                tf.keras.layers.Dense(output_size)
            ])
        else:
            raise ValueError(f"Unsupported TensorFlow model type: {model_type}")

    def _generate_result_hash(self, result: Any) -> str:
        """
        Generate a cryptographic hash of the task result
        """
        result_str = json.dumps(result, sort_keys=True)
        return hashlib.sha256(result_str.encode()).hexdigest()

    def validate_task_result(self, original_result: Dict[str, Any], processed_result: Dict[str, Any]) -> bool:
        """
        Validate the processed task result against original expectations
        """
        if processed_result['status'] != 'completed':
            return False

        # Compare result hash
        if 'result_hash' not in processed_result:
            return False

        # Additional validation logic can be added here
        return True
